+++ src/spat.zig
@@ -1,1 +1,1 @@
-const std = @import("std");
+pub const argu = @import("argu.zig");